Move EventQueue and implementations to internal package. Move TimestampConverter to internal implementation. Move RandomHandler to trace/internal.
