+++ core_impl/src/main/java/io/opencensus/trace/SpanFactoryImpl.java
@@ -16,6 +16,7 @@ package io.opencensus.trace;
 import static com.google.common.base.Preconditions.checkNotNull;
 
 import io.opencensus.common.Clock;
+import io.opencensus.internal.TimestampConverter;
 import io.opencensus.trace.Span.Options;
 import io.opencensus.trace.base.Link;
 import io.opencensus.trace.base.Link.Type;
@@ -24,6 +25,7 @@ import io.opencensus.trace.base.TraceId;
 import io.opencensus.trace.base.TraceOptions;
 import io.opencensus.trace.config.TraceConfig;
 import io.opencensus.trace.config.TraceParams;
+import io.opencensus.trace.internal.RandomHandler;
 import java.util.EnumSet;
 import java.util.List;
 import java.util.Random;
