+++ core_impl/src/test/java/com/google/instrumentation/stats/StatsContextTest.java
@@ -19,7 +19,7 @@ import static org.junit.Assert.fail;
 import com.google.common.collect.Collections2;
 import com.google.common.testing.EqualsTester;
 import io.opencensus.common.Function;
-import io.opencensus.common.SimpleEventQueue;
+import io.opencensus.internal.SimpleEventQueue;
 import io.opencensus.internal.TestClock;
 import io.opencensus.internal.VarInt;
 import com.google.instrumentation.stats.View.DistributionView;
