+++ core_impl/src/test/java/io/opencensus/trace/SpanImplTest.java
@@ -18,6 +18,7 @@ import static com.google.common.truth.Truth.assertThat;
 import io.opencensus.common.Duration;
 import io.opencensus.common.Timestamp;
 import io.opencensus.internal.TestClock;
+import io.opencensus.internal.TimestampConverter;
 import io.opencensus.trace.Span.Options;
 import io.opencensus.trace.SpanImpl.StartEndHandler;
 import io.opencensus.trace.base.Annotation;
