+++ core_impl/src/test/java/io/opencensus/trace/TraceComponentImplBaseTest.java
@@ -16,8 +16,8 @@ package io.opencensus.trace;
 import static com.google.common.truth.Truth.assertThat;
 
 import io.opencensus.common.MillisClock;
-import io.opencensus.common.SimpleEventQueue;
-import io.opencensus.trace.RandomHandler.SecureRandomHandler;
+import io.opencensus.internal.SimpleEventQueue;
+import io.opencensus.trace.internal.RandomHandler.SecureRandomHandler;
 import io.opencensus.trace.propagation.PropagationComponentImpl;
 import org.junit.Test;
 import org.junit.runner.RunWith;
